add restaurant entity score attributes. create review class

+++ src/main/java/com/codecademy/dinningReview/model/Rating.java
@@ -1,11 +1,11 @@
 package com.codecademy.dinningReview.model;
 
 public enum Rating {
-    ONE(1), TWO(2), THREE(3), FOUR(4), FIVE(5);
+    ONE(1.0), TWO(2.0), THREE(3.0), FOUR(4.0), FIVE(5.0);
 
-    public final int stars;
+    public final Double stars;
 
-    Rating(int stars) {
+    Rating(Double stars) {
         this.stars = stars;
     }
